refactoring uploadUtils for better naming
