+++ app/javascript/components/upload/ClusteringFileForm.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import Select from 'lib/InstrumentedSelect'
 import FileUploadControl, { FileTypeExtensions } from './FileUploadControl'
-import { TextFormField, SavingOverlay, SaveDeleteButtons } from './uploadUtils'
+import { TextFormField, SavingOverlay, SaveDeleteButtons } from './form-components'
 
 /** renders a form for editing/uploading a single cluster file */
 export default function ClusteringFileForm({
@@ -21,6 +21,7 @@ export default function ClusteringFileForm({
     <div className="col-md-12">
       <form id={`clusterForm-${file._id}`}
         className="form-terra"
+        onSubmit={e => e.preventDefault()}
         acceptCharset="UTF-8">
         <FileUploadControl
           handleSaveResponse={handleSaveResponse}
