+++ app/javascript/components/upload/CoordinateLabelStep.js
@@ -1,6 +1,7 @@
 import React, { useEffect } from 'react'
 
 import CoordinateLabelFileForm from './CoordinateLabelFileForm'
+import { AddFileButton } from './form-components'
 
 const DEFAULT_NEW_LABEL_FILE = {
   file_type: 'Coordinate Labels',
@@ -78,7 +79,6 @@ function CoordinateLabelForm({
         </div>
       </div>
     </div>
-
     { coordinateFiles.map(file => {
       return <CoordinateLabelFileForm
         key={file._id}
@@ -90,8 +90,6 @@ function CoordinateLabelForm({
         associatedClusterFileOptions={associatedClusterFileOptions}
         updateCorrespondingClusters={updateCorrespondingClusters}/>
     })}
-    <div className="row top-margin">
-      <button className="btn btn-secondary action" onClick={() => addNewFile(DEFAULT_NEW_LABEL_FILE)}><span className="fas fa-plus"></span> Add File</button>
-    </div>
+    <AddFileButton addNewFile={addNewFile} newFileTemplate={DEFAULT_NEW_LABEL_FILE}/>
   </div>
 }
