+++ app/javascript/components/upload/ExpressionFileForm.js
@@ -4,7 +4,7 @@ import _kebabCase from 'lodash/kebabCase'
 import Select from 'lib/InstrumentedSelect'
 import MTXBundledFilesForm from './MTXBundledFilesForm'
 import FileUploadControl, { FileTypeExtensions } from './FileUploadControl'
-import { TextFormField, SavingOverlay, SaveDeleteButtons } from './uploadUtils'
+import { TextFormField, SavingOverlay, SaveDeleteButtons } from './form-components'
 
 /** renders a form for editing/uploading an expression file (raw or processed) and any bundle children */
 export default function ExpressionFileForm({
@@ -26,6 +26,7 @@ export default function ExpressionFileForm({
     <div className="col-md-12">
       <form id={`clusterForm-${file._id}`}
         className="form-terra"
+        onSubmit={e => e.preventDefault()}
         acceptCharset="UTF-8">
 
         <FileUploadControl
